requirements.txt
streamlit
numpy
pandas
scikit-learn
scikit-image
Pillow
joblib
gdown